Add standard library classes to the semantic analyzer
